web_service/fastapi/pyproject.toml
[tool.poetry]
name = "openeo-fastapi-service"
version = "0.1.0"
description = ""
authors = ["Tomas Kliment <tomas@klimeto.com>", 'Pieter Kempeneers <pieter.kempeneers@ec.europ.eu>']

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.61.1"
uvicorn = "^0.11.8"
fastapi_utils = "^0.2.1"
alembic = "^1.4.3"
sqlalchemy = "^1.3.19"
geoalchemy2 = "^0.8.4"
sqlalchemy_utils = "^0.36.8"
psycopg2-binary = "^2.8.6"
sqlalchemy_mixins = "^1.2.1"
shapely = "^1.7.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
